backend/pyproject.toml
[project]
name = "new-vitriona-backend"
version = "0.1.0"
description = "FastAPI rewrite of vitriona_construct"
authors = [
    { name = "OpenAI Codex" }
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.127.0,<0.128.0",
    "uvicorn[standard]>=0.40.0,<0.41.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "sqlalchemy[asyncio]>=2.0.45,<3.0.0",
    "alembic>=1.17.2,<2.0.0",
    "psycopg[binary]>=3.3.2,<4.0.0",
    "asyncpg>=0.31.0,<0.32.0",
    "redis>=7.1.0,<8.0.0",
    "python-multipart>=0.0.20,<0.0.21",
    "itsdangerous>=2.2.0,<3.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "bcrypt>=3.2.2,<4.0.1",
    "httpx>=0.28.1,<0.29.0",
    "jinja2>=3.1.6,<4.0.0",
    "email-validator>=2.3.0,<3.0.0",
    "dependency-injector>=4.48.3,<5.0.0",
    "python-jose[cryptography]>=3.5.0,<4.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "src" },
]

[dependency-groups]
dev = [
    "pytest>=9.0.2,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "black>=25.12.0,<26.0.0",
    "isort>=7.0.0,<8.0.0",
]
